Only search for 50 ordinary shareholders
For companies where Creditsafe returns a large number of ordinary (i.e
non-PSC or non-officer shareholders) we can end up doing a lot of extra
searches against their API.

In some cases this can cause the report to take more than 120s to be
completed, which results in Gunicorn killing the worker process.

Here we add a 'should_search' flag to shareholder objects and set it to
false on all shareholders after the 50th (ordered by total percentage
shareholding).

When doing the lookup, we check this flag if the associated entity is
neither a PSC nor an Officer, and skip the search if appropriate.