Add Matches count field and correct use of SSN field
The matches field was not included in the match structure, leading to an error
when fed into the monolith. This was amended. Additionally, the SSN field was
not used correctly (it didn't drill down into the actual country SSN field) so
it was not being passed to the LexisNexis API correctly.