Pass `is_cta` from monolith rather than URL
Instead of including the target URLs in the monolith config and sending
them to the integration with every request, we include these in the
integration code and select between them with a boolean parameter.

`is_cta` indicates that the request should be processed against the
Equifax Customer Test Area instead of the production environment.