Check if company exists before adding to creditsafe portfolio
If we try to add a company to a creditsafe portfolio which is already in
the portfolio, creditsafe will return an error. We only want to try and
add a company to a portfolio if it doesn't exist already.

This change hits the companies endpoint to check if the company exists
in a portfolio. If this gives a 404 or a 400 with a specific error
(their documentation says it will return a 404 but it actually returns a
400 with PortfolioCompany not found error message...) then add the company to
the portfolio.