Upgrade all integration containers
In line with the ongoing security cleanup work triggered by Prisma/Twistlock,
all integrations containers are being rebuilt to incorporate the latest security
updates.

I've also taken this time to unify all integrations to be running on the
official Alpine-based Python images (latest 3.7 for all, except WorldCheck
which was specified for 3.6 and DJW API which uses 3.8, so use those).