USe default country if missing from creditsafe address
Creditsafe does not always send the country as part of the address, and country is a required field for us, hence we can't add the address to our system without a country.

Whenever the country of address is not returned we will assume it is the same as the country of incorporation and therefore be able to display in the check result
The ideal solution would be to display the address regardless of whether the country is available or not but this would require fundamental changes to how our product work which increase the work and associated risk.