Always use UTC epochs in HTTP Signature auth
Passing a time struct parsed from the request into `time.mktime` will
result in an epoch time using the system timezone.

Since we use `time.time()` for the other side of this comparison (which
is always UTC), we need to make sure both epoch timestamps are UTC
before doing the check.