Remove removing post code and street address
On further investigation, the post code and street address is simply not
allowed to be filled in PassFort testing account as Iovation rejects
those fields from UK based companies (us). So I have added those fields
back. In the future we will need to have this as a configuration option
if any of our institutions from the US/UK use iovation