Expose if test environment was used
In device fraud detection add if test environment was used. Logic for it
being there is that it exposes if the fraud detection is genuine or just
some test info.

This information is needed so the links to iovation can be shown
conditionally on the FE, i.e if test environment link to
ci-admin.iovation rather than admin.iovation