Limit number of Creditsafe shareholders to 50
Restrict the number of shareholder associates returned in a Creditsafe
report to 50. This is a bit of an arbitrary number, but we have to draw
the line somewhere as we're seeing reports with more than 300
associates.

Also limit the number of 'additional PSCs' we try and process, although
this list is usually small or empty.