Use BvD ID to generate associate ID if present
Associate ID ought to be stable across multiple checks so that we can
match monitoring results to previous check results. Unfortunately, BvD
don't attach a stable ID to every field in their response.

We seem to get a BvD ID for:
- Every shareholder entity
- Every beneficial owner entity
- Company officers

Additionally we get a UCI field which may change over time.

Here's what BvD said when I asked about using UCI to dedup:

> The UCI is used to identify individuals available in the Contacts
> database and it can be often changed due to the fact that we are
> merging different data. So if you would like to use it to create a
> single profile, please create a weekly check to see whether the UCI
> got updated.

For now the least bad approach seems to be merging associates by
UCI (which is unstable but seems to be the field that is most
consistently filled across officers) and using the BvD ID to generate
the associate ID (falling back to full name if it's not present).