Search by associate country not parent company
For company associates, the integration searches for it according to
name and the country of incorporation of the parent. This doesn't work
if the associate is in another country and can return the wrong result
if their is another company with the same name in the country of the
parent.

Fix this by looking up the country of the associate in extended group
structure and using that to search.

The second change is if search data can not be found for an associate,
and we know the country from extended group structure, then use that
country.