new company search fields APPS-333
Added support for the new search response fields to this reference
integration. Also added some demo data to MANY_HITS.json.

Since these new fields are optional the first company in the demo data
has all the fields but the subsequent companies do not to ensure that we
allow for that.

Also added a test for the search to ensure it processes the fields
correctly.