New monitored polling reference API implementation
Based on existing Orbis polling implementation, added a reference
implementation that uses the same demo data as those used by the
synchronous reference implementation.

To mimic a polling system, saves the demo data in a temporary file and
serves this as the check result. Added a private API that allows for
updating the check result to enable triggering of new snapshots in the
bridge.