Change monitored polling endpoints for company data reference integration
We are transitioning monitored polling in the bridge to remove reference
number from the endpoint URL. This is because it is being duplicated in
the request body (where it will remain) and reference numbers may
contain URL reserved characters which adds complications with encoding.

The reference number has been removed from monitored_check/ready
and monitored_checks/poll to match where the bridge is sending
the requests now.